+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Client Side API:
 * Removed `wc/compat` RequireJS loader plugin.
 * Removed `wc/template`, use native template options like JS template strings and HTML template.
 * Removed `wc/array/toArray`, use native `Array.from`.
+* Removed `wc/global`, use native `globalThis` instead.
 * Nerfed `wc/has`, it is no longer poweful and backed by dojo (we don't do this as much anymore).
 * Removed RequireJS (`window.require` and `window.define`) - use native modules.